app.js: determine locale by browser settings.
In your browser you can configure a list of languages that you can understood and you can set priority for them.
In Chrome chrome://settings/?search=Language
Then we can see the list of locales via navigator.languages property or Accept-Language header. It looks like ['de', 'en-US', 'en'] i.e. it can contain only lang code or lang code and country.

So here is added a short language determiner that on loading sets a locale if it requested by a a user and is supported.